pull out decay code download into separate script called by cmake
